+++ Plugins/GW2_UI/GW2_UI.lua
@@ -505,11 +505,11 @@ local function SetAchievement(self, achievement, refresh)
     local state = achievement:GetObtainableState();
 
     local texture;
-    if state and (state == false or state == "Past") then
+    if state == "Past" then
         texture = "Interface/AddOns/GW2_UI/textures/uistuff/achievementcompletebgred";
-    elseif state and state == "Current" then
+    elseif state == "Current" then
         texture = media .. "GW2_UI-achievementcompletebggreen";
-    elseif state and state == "Future" then
+    elseif state == "Future" then
         texture = media .. "GW2_UI-achievementcompletebgyellow";
     else
         if self.accountWide then